Copy over tests added from #13333
